Creating reusable component for explore page. Fixing component names in bookmark page
